+++ packages/cosmology/src/clients/validator.js
@@ -1,63 +1,14 @@
 import { RestClient } from './rest';
 import autobind from 'class-autobind';
 
-/**
- *
- *
- * @typedef {{
- *  denom: string;
- *  exponent: number;
- *  name: string;
- *  symbol: string;
- *  price: number;
- *  price_24h_chainge: number;
- *  liquidity: number;
- *  liquidity_24h_change: number;
- *  volume_24h: number;
- *  volume_24h_change: number;
- * }} ValidatorToken
- *
- * @typedef {{
- *  price: number;
- * }} ValidatorTokenPrice
- *
- * @typedef {{
- * start_date: string;
- * denom: string;
- * symbol: string;
- * apr_1d: number;
- * apr_7d: number;
- * apr_14d: number;
- * }} ValidatorCoinApr
- *
- * @typedef {{
- *  pool_id: number;
- *  apr_list: ValidatorCoinApr[]
- * }} ValidatorPoolApr
- *
- *
- * interestingly this one has "price" which I think is the
- * price of the LP token? TODO: verify...
- *
- * @typedef {{
- *  pool_address: string;
- *  pool_id: string;
- *  base_name: string;
- *  base_symbol: string;
- *  base_address: string;
- *  quote_name: string;
- *  quote_symbol: string;
- *  quote_address: string;
- *  price: number;
- *  base_volume_24h: number;
- *  quote_volume_24h: number;
- *  volume_24h: number;
- *  volume_7d: number;
- *  liquidity: number;
- *  liquidity_atom: number;
- * }} ValidatorPair
- *
- */
+import {
+  ValidatorToken,
+  CoinSymbol,
+  ValidatorPoolApr,
+  ValidatorPair,
+  ValidatorTokenPrice
+} from '../types';
+
 //  https://api-osmosis.imperator.co/swagger/#/
 export class OsmosisValidatorClient extends RestClient {
   constructor({ url = 'https://api-osmosis.imperator.co/' } = {}) {
@@ -80,7 +31,7 @@ export class OsmosisValidatorClient extends RestClient {
   }
 
   /**
-   * @param {string} symbol
+   * @param {CoinSymbol} symbol
    * @returns {Promise<ValidatorToken[]>}
    */
   async getToken(symbol) {
@@ -89,7 +40,7 @@ export class OsmosisValidatorClient extends RestClient {
   }
 
   /**
-   * @param {string} symbol
+   * @param {CoinSymbol} symbol
    * @returns {Promise<ValidatorTokenPrice[]>}
    */
   async getTokenPrice(symbol) {
@@ -106,7 +57,7 @@ export class OsmosisValidatorClient extends RestClient {
   }
 
   /**
-   * @param {string} symbol
+   * @param {CoinSymbol} symbol
    * @returns {Promise<ValidatorPoolApr[]>}
    */
   async getPoolApr(symbol) {
@@ -115,11 +66,6 @@ export class OsmosisValidatorClient extends RestClient {
   }
 
   /**
-   * TODO: incorrect type (export in index.d.ts)
-   * {
-   * updated_at: number
-   * data:ValidatorPair[]
-   * }
    * @returns {Promise<ValidatorPair[]>}
    */
   async getPairsSummary() {