+++ packages/cosmology/src/commands/lock.js
@@ -4,6 +4,11 @@ import { osmoRestClient } from '../utils';
 import { getSigningOsmosisClient } from '../messages/utils';
 import { messages } from '../messages/messages';
 import { signAndBroadcast } from '../messages/utils';
+import {
+  printSwap,
+  printSwapForPoolAllocation,
+  printOsmoTransactionResponse
+} from '../utils/print';
 
 const osmoChainConfig = chains.find((el) => el.chain_name === 'osmosis');
 const rpcEndpoint = osmoChainConfig.apis.rpc[0].address;
@@ -58,14 +63,14 @@ export default async (argv) => {
   switch (duration) {
     case 1:
     case '1':
-      duration = '86400000';
+      duration = '86400';
       break;
     case 7:
     case '7':
-      duration = '604800000';
+      duration = '604800';
       break;
     default:
-      duration = '1209600000';
+      duration = '1209600';
       break;
   }
 
@@ -78,7 +83,9 @@ export default async (argv) => {
     duration
   });
 
-  console.log(JSON.stringify(msg, null, 2));
+  if (argv.verbose) {
+    console.log(JSON.stringify(msg, null, 2));
+  }
 
   const accounts = await signer.getAccounts();
   const osmoAddress = accounts[0].address;
@@ -96,11 +103,5 @@ export default async (argv) => {
     memo: ''
   });
 
-  if (res.transactionHash) {
-    console.log(`tx hash ${res.transactionHash}`);
-  } else {
-    console.log('no tx found!');
-  }
-  console.log('\n\n\n\n\ntx');
-  console.log(res);
+  printOsmoTransactionResponse(res);
 };