+++ packages/cosmology/src/index.js
@@ -3,4 +3,3 @@ export * from './clients';
 export * from './utils/osmo';
 export * from './utils/chain';
 export * from './messages';
-export * from './model';