+++ packages/cosmology/src/messages/messages.js
@@ -6,6 +6,11 @@ import { generateOsmoMessage as osmo } from './utils';
  * tokenOutDenom:string;
  * }} Route
  *
+ * @typedef {{
+ * amount:string;
+ * denom:string;
+ * }} Coin
+ *
  */
 
 export const messages = {
@@ -17,7 +22,7 @@ export const messages = {
    * @param {string} param0.sender
    * @param {string} param0.poolId
    * @param {string} param0.shareOutAmount
-   * @param {import('../utils').Coin[]} param0.tokenInMaxs
+   * @param {Coin[]} param0.tokenInMaxs
    */
   joinPool: ({ sender, poolId, shareOutAmount, tokenInMaxs }) =>
     osmo('joinPool', {
@@ -30,7 +35,7 @@ export const messages = {
    * @param {object} param0
    * @param {string} param0.sender
    * @param {string} param0.poolId
-   * @param {import('../utils').Coin} param0.tokenIn
+   * @param {Coin} param0.tokenIn
    * @param {string} param0.shareOutMinAmount
    */
   joinSwapExternAmountIn: ({ sender, poolId, tokenIn, shareOutMinAmount }) =>
@@ -47,7 +52,7 @@ export const messages = {
    * @param {object} param0
    * @param {string} param0.sender
    * @param {Route[]} param0.routes
-   * @param {import('../utils').Coin} param0.tokenIn
+   * @param {Coin} param0.tokenIn
    * @param {string} param0.tokenOutMinAmount
    */
   swapExactAmountIn: ({ sender, routes, tokenIn, tokenOutMinAmount }) =>
@@ -64,7 +69,7 @@ export const messages = {
    * @param {object} param0
    * @param {string} param0.owner
    * @param {string} param0.duration
-   * @param {import('../utils').Coin[]} param0.coins
+   * @param {Coin[]} param0.coins
    */
   lockTokens: ({ owner, duration, coins }) =>
     osmo('lockTokens', { owner, duration, coins }),