+++ packages/registry/CHANGELOG.md
@@ -3,7 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<<<<<<< HEAD
 ## [0.0.6](https://github.com/pyramation/cosmology/compare/@cosmology/cosmos-registry@0.0.5...@cosmology/cosmos-registry@0.0.6) (2022-04-01)
+=======
+## [0.0.9](https://github.com/cosmology-finance/cosmology/compare/@cosmology/cosmos-registry@0.0.8...@cosmology/cosmos-registry@0.0.9) (2022-04-05)
+>>>>>>> d3e27860a53a2eb1dee466be442b0389c31eff91
 
 **Note:** Version bump only for package @cosmology/cosmos-registry
 
@@ -11,7 +15,11 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+<<<<<<< HEAD
 ## [0.0.5](https://github.com/pyramation/cosmology/compare/@cosmology/cosmos-registry@0.0.4...@cosmology/cosmos-registry@0.0.5) (2022-04-01)
+=======
+## [0.0.8](https://github.com/cosmology-finance/cosmology/compare/@cosmology/cosmos-registry@0.0.7...@cosmology/cosmos-registry@0.0.8) (2022-04-05)
+>>>>>>> d3e27860a53a2eb1dee466be442b0389c31eff91
 
 **Note:** Version bump only for package @cosmology/cosmos-registry
 
@@ -19,7 +27,35 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+<<<<<<< HEAD
 ## [0.0.4](https://github.com/pyramation/cosmology/compare/@cosmology/cosmos-registry@0.0.3...@cosmology/cosmos-registry@0.0.4) (2022-03-16)
+=======
+## [0.0.7](https://github.com/cosmology-finance/cosmology/compare/@cosmology/cosmos-registry@0.0.6...@cosmology/cosmos-registry@0.0.7) (2022-04-01)
+
+**Note:** Version bump only for package @cosmology/cosmos-registry
+
+
+
+
+
+## [0.0.6](https://github.com/cosmology-finance/cosmology/compare/@cosmology/cosmos-registry@0.0.5...@cosmology/cosmos-registry@0.0.6) (2022-04-01)
+
+**Note:** Version bump only for package @cosmology/cosmos-registry
+
+
+
+
+
+## [0.0.5](https://github.com/cosmology-finance/cosmology/compare/@cosmology/cosmos-registry@0.0.4...@cosmology/cosmos-registry@0.0.5) (2022-04-01)
+
+**Note:** Version bump only for package @cosmology/cosmos-registry
+
+
+
+
+
+## [0.0.4](https://github.com/cosmology-finance/cosmology/compare/@cosmology/cosmos-registry@0.0.3...@cosmology/cosmos-registry@0.0.4) (2022-03-16)
+>>>>>>> d3e27860a53a2eb1dee466be442b0389c31eff91
 
 **Note:** Version bump only for package @cosmology/cosmos-registry